+++ src/plugins/applyPrettier.ts
@@ -1,11 +1,9 @@
 import { type FileHandle, open } from 'node:fs/promises'
-
 import type { Linter } from 'eslint'
-
-import type { PartialProfileConfig } from 'src/types/interfaces'
-
-import { cwd } from '../constants'
-import type { ArrayOption } from '../types'
+import type { PartialProfileConfig, ShinyConfig } from 'src/types/interfaces'
+import type { ArrayOption } from '../types/types'
+import { join } from 'node:path'
+import fileToJson from 'src/utils/fileToJson'
 
 const prettierRuleDict: Record<string, string> = {
     arrowParens: 'arrow-parens',
@@ -24,7 +22,7 @@ const maxLenDict: Record<string, string> = {
     tabWidth: 'tabWidth'
 }
 
-const numericalRules = new Set(['printWidth', 'tabWidth'])
+const numericalRules = new Set(['printWidth', 'tabWidth', 'useTabs'])
 const banWords = new Set(['avoid', 'false', 'none', 'preserve'])
 const ignore = new Set([
     'plugins',
@@ -36,7 +34,6 @@ const ignore = new Set([
     'jsxBracketSameLine',
     'jsxSingleQuote',
     'singleAttributePerLine',
-    'useTabs',
     'vueIndentScriptAndStyle',
     'htmlWhitespaceSensitivity',
     'proseWrap',
@@ -49,14 +46,60 @@ const ignore = new Set([
 
 const jsPlugin = '@stylistic/js'
 const tsPlugin = '@stylistic/ts'
-const measureRule = `${jsPlugin}/max-len`
+const maxLenRule = `${jsPlugin}/max-len`
+const indentRule = `${tsPlugin}/indent`
 
-function handleMeasurements(rules: Linter.RulesRecord, rule: string, prettierValue: number): void {
-    let value: ArrayOption | undefined = rules[measureRule] as ArrayOption | undefined
-    // init the value
-    if (!value) value = rules[measureRule] = [2, {}]
-    // get the rule from the prettier - max-len dict
-    value[1][maxLenDict[rule]] = prettierValue
+function setIndentValue(rule: any, useTabs: boolean, prettierValue: boolean | number, extraOptions?: Record<string, unknown>): any {
+    if (rule) return rule
+    const value = [2, useTabs && prettierValue ? 'tab' : prettierValue || 4]
+    if (extraOptions) value.push(extraOptions)
+    return value
+}
+
+function handleMeasurements(opts: ShinyConfig, rules: Linter.RulesRecord, rule: string, prettierValue: boolean | number): void {
+    const isTabWidth = rule === 'tabWidth'
+    if (rule === 'printWidth' || isTabWidth) {
+        let value: ArrayOption | undefined = rules[maxLenRule] as ArrayOption | undefined
+        if (!value) value = rules[maxLenRule] = [2, {}]
+        value[1][maxLenDict[rule]] = prettierValue
+    }
+    const isUseTabs = rule === 'useTabs'
+    if ((isUseTabs || isTabWidth) && opts.indent) {
+        let value: any = rules[indentRule]
+        if (!value) value = rules[indentRule] = [2, {}]
+        if (isUseTabs && prettierValue) value[1] = 'tab'
+        if (isTabWidth && typeof prettierValue === 'number') {
+            value[1] = value[1] === 'tab' ? value[1] : prettierValue
+            // options
+            if (!value[2]) {
+                const halfIndent = Math.floor(prettierValue / 2)
+                value[2] = {
+                    ArrayExpression: halfIndent,
+                    CallExpression: halfIndent,
+                    FunctionExpression: {
+                        body: prettierValue,
+                        parameters: halfIndent
+                    },
+                    ImportDeclaration: halfIndent,
+                    MemberExpression: halfIndent,
+                    ObjectExpression: halfIndent,
+                    StaticBlock: prettierValue,
+                    SwitchCase: halfIndent,
+                    VariableDeclarator: 'first',
+                    flatTernaryExpressions: false,
+                    ignoreComments: false,
+                    offsetTernaryExpressions: false
+                }
+            }
+        }
+        if (opts.configs.includes('vue')) rules['vue/html-indent'] = setIndentValue(rules['vue/html-indent'], isUseTabs, prettierValue)
+        if (opts.configs.includes('react')) {
+            rules['@stylistic/jsx/jsx-indent'] = setIndentValue(rules['@stylistic/jsx/jsx-indent'], isUseTabs, prettierValue, {
+                checkAttributes: true,
+                indentLogicalExpressions: true
+            })
+        }
+    }
 }
 
 function applyAdditionalRules(rules: Linter.RulesRecord, usedPlugin: string, rule: string, isFalseValue: boolean): void {
@@ -109,20 +152,20 @@ function mapToEslint(rules: Linter.RulesRecord, rule: string, value: boolean | s
     applyAdditionalRules(rules, usedPlugin, convertedRule, isFalseValue)
 }
 
-export default async function applyPrettier(): Promise<PartialProfileConfig> {
+export default async function applyPrettier(opts: ShinyConfig): Promise<PartialProfileConfig> {
     let file: FileHandle
     const rules: Linter.RulesRecord = {}
     try {
-        file = await open(`${cwd}/.prettierrc`, 'r')
+        file = await open(join(opts.root, '.prettierrc'), 'r')
     } catch {
         return { name: 'prettier-apply', rules: [] }
     }
 
-    const json = JSON.parse((await file.readFile()).toString())
+    const json = await fileToJson(file)
     for (const key of Object.keys(json)) {
         if (!ignore.has(key)) {
             // Handle numerical rules. Those are measurement rules
-            if (numericalRules.has(key)) handleMeasurements(rules, key, json[key])
+            if (numericalRules.has(key)) handleMeasurements(opts, rules, key, json[key])
             else mapToEslint(rules, key, json[key])
         }
     }