+++ src/plugins/findTSConfigs.ts
@@ -1,10 +1,9 @@
 import { fdir } from 'fdir'
 
-import { cwd } from 'src/constants'
-import type { PartialProfileConfig } from 'src/types/interfaces'
+import type { PartialProfileConfig, ShinyConfig } from 'src/types/interfaces'
 
-export default async function findTSConfigs(): Promise<PartialProfileConfig> {
-    const api = new fdir().withFullPaths().withMaxDepth(1).crawl(cwd)
+export default async function findTSConfigs(opts: ShinyConfig): Promise<PartialProfileConfig> {
+    const api = new fdir().withFullPaths().withMaxDepth(1).crawl(opts.root)
     const files = await api.withPromise()
     const length = files.length
     const tsconfigFiles: string[] = []
@@ -18,6 +17,7 @@ export default async function findTSConfigs(): Promise<PartialProfileConfig> {
             parserOptions: {
                 project: tsconfigFiles
             }
-        }
+        },
+        name: 'tsconfig-resolve'
     }
 }