+++ src/plugins/patchVSCode.ts
@@ -1,7 +1,8 @@
 import { existsSync } from 'node:fs'
 import { mkdir, open, writeFile } from 'node:fs/promises'
-
-import { cwd } from 'src/constants'
+import { join } from 'node:path'
+import type { ShinyConfig } from 'src/types/interfaces'
+import fileToJson from 'src/utils/fileToJson'
 
 const VSCodePatch = {
     // Auto fix
@@ -41,20 +42,24 @@ const VSCodePatch = {
 
 const keys = Object.keys(VSCodePatch)
 
-export default async function patchVSCode(): Promise<void> {
-    const vsCodeFolderPath = `${cwd}/.vscode`
-    const settingsPath = `${vsCodeFolderPath}/settings.json`
+export default async function patchVSCode(opts: ShinyConfig): Promise<void> {
+    const vsCodeFolderPath = join(opts.root, '.vscode')
+    const settingsPath = join(vsCodeFolderPath, 'settings.json')
     if (!existsSync(vsCodeFolderPath)) await mkdir(vsCodeFolderPath)
     if (!existsSync(settingsPath)) await writeFile(settingsPath, JSON.stringify(VSCodePatch), 'utf8')
     const file = await open(settingsPath, 'r+')
-    const settings = JSON.parse((await file.readFile()).toString())
+    const settings = await fileToJson(file)
     let shouldWrite = true
     for (const key of keys) {
+        // only write when there are no eslint keys inside the settings.json
         if (settings[key]) {
             shouldWrite = false
             break
         }
     }
-    if (shouldWrite) await file.writeFile(JSON.stringify(Object.assign(settings, VSCodePatch)))
+    if (shouldWrite) {
+        const buffer = Buffer.from(JSON.stringify(Object.assign(settings, VSCodePatch)))
+        await file.write(buffer, 0, buffer.byteLength, 0)
+    }
     await file.close()
 }