+++ src/profiles/vue.ts
@@ -13,6 +13,22 @@ export const config: PartialProfileConfig = {
     extends: ['web'],
     files: ['**/*.vue'],
     languageOptions: {
+        globals: {
+            computed: 'readonly',
+            defineEmits: 'readonly',
+            defineExpose: 'readonly',
+            defineProps: 'readonly',
+            onMounted: 'readonly',
+            onUnmounted: 'readonly',
+            reactive: 'readonly',
+            ref: 'readonly',
+            shallowReactive: 'readonly',
+            shallowRef: 'readonly',
+            toRef: 'readonly',
+            toRefs: 'readonly',
+            watch: 'readonly',
+            watchEffect: 'readonly'
+        },
         parser: vueParser,
         parserOptions: {
             ecmaFeatures: {
@@ -50,6 +66,7 @@ export const config: PartialProfileConfig = {
         vue.configs['vue3-strongly-recommended'],
         vue.configs['vue3-essential'],
         vueCss.configs['vue3-recommended'],
+        vueAccess.configs['flat/recommended'][1],
         ban(GeneralBanList, ['vue']),
         replace(StyleVueReplaceList, ['@stylistic/ts'], ['vue']),
         {